tests/wpt/web-platform-tests/preload/single_download_preload.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    var t = async_test('Makes sure that preloaded resources are not downloaded again when used');
</script>
<link rel=preload href="resources/dummy.js" as=script>
<link rel=preload href="resources/dummy.css" as=style>
<link rel=preload href="resources/square.png" as=image>
<link rel=preload href="resources/square.png?background" as=image>
<link rel=preload href="/fonts/CanvasTest.ttf" as=font crossorigin>
<link rel=preload href="/media/white.mp4" as=media>
<link rel=preload href="/media/sound_5.oga" as=media>
<link rel=preload href="/media/foo.vtt" as=media>
<link rel=preload href="resources/dummy.xml?foo=bar" as=foobarxmlthing>
<link rel=preload href="resources/dummy.xml">
<body>
<script src="resources/dummy.js?pipe=trickle(d5)"></script>
<style>
    #background {
        width: 200px;
        height: 200px;
        background-image: url(resources/square.png?background);
    }
    @font-face {
      font-family:ahem;
      src: url(/fonts/CanvasTest.ttf);
    }
    span { font-family: ahem, Arial; }
</style>
<link rel="stylesheet" href="resources/dummy.css">
<script src="resources/dummy.js"></script>
<div id="background"></div>
<img src="resources/square.png">
<video src="/media/white.mp4">
    <track kind=subtitles src="/media/foo.vtt" srclang=en>
</video>
<audio src="/media/sound_5.oga"></audio>
<script>
    var xhr = new XMLHttpRequest();
    xhr.open("GET", "resources/dummy.xml");
    xhr.send();

    window.addEventListener("load", t.step_func(function() {
        // Audio and video show 2 extra requests as the main request is followed by a range request
        assert_equals(performance.getEntriesByType("resource").length, 14);
        t.done();
    }));
</script>
<span>PASS - this text is here just so that the browser will download the font.</span>
