third_party/WebKit/LayoutTests/external/wpt/preload/single_download_preload-expected.txt

CONSOLE WARNING: line 15: <link rel=preload> must have a valid `as` value
This is a testharness.js-based test.
FAIL Makes sure that preloaded resources are not downloaded again when used assert_equals: expected 14 but got 0
Harness: the test ran to completion.
